requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
astroid==3.0.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
crosstalk-correction==1.0.5
dash==2.10.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
executing==1.2.0
Flask==2.2.5
Flask-Cors==4.0.0
frozenlist==1.4.0
gql==3.5.0b5
graphql-core==3.3.0a3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.25.2
ipython==8.15.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.25.2
openai==1.3.2
packaging==23.1
pandas==2.1.0
pandas-stubs==2.0.3.230814
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.15.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
pycodestyle==2.11.1
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pylint==3.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-toolbelt==0.10.1
six==1.16.0
sniffio==1.3.0
specklepy==2.17.11
stack-data==0.6.2
stringcase==1.2.0
tenacity==8.2.2
tomlkit==0.12.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
types-pytz==2023.3.1.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
#urllib3==2.1.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
